refactor: rename ast.lua to query.lua
